Create 3207 Lab 2 Basic Skeleton after some tinkering
Redesigned "quit" so that it's a command as opposed to a a condition tested by the shell's while loop. This allows me to keep the shell running while not feof, so that I don't need a separate version of the while loop for batch mode AND don't need a super-complicated condition.