Create 3207 Lab 3 Condition Variable Practice
So after all that Bryant & O'Hallaron (and a little 3 Easy Pieces, because B&O doesn't really cover condition variables) I made this to make sure that I understood how condition variables worked. Made a toy producer-consumer problem, with an int that could be incremented up through MAX_ITEMS; if the int was 0, that was the equivalent of an empty data structure, whereas if it was maxed out, it was like a filled-up data structure. The program checks whether the int ever gets smaller than 0 or bigger than MAX_Items; sure enough, without the condition variables, it does, but with them, it doesn't.